calendar dialog fixes

+++ backend/src/Controller/Finance/Calendar/CalendarController.php
@@ -11,12 +11,14 @@ use App\Entity\User;
 use App\Repository\CalendarRepository;
 use App\Repository\ExpenseRepository;
 use App\Request\Calendar\CreateCalendarRequest;
+use App\Request\Calendar\UpdateCalendarRequest;
 use App\Response\Statement\ExpenseListResponse;
 use App\Service\BalanceCalculatorService;
 use DateTime;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Http\Attribute\CurrentUser;
 
 #[Route('api/calendar', name: 'calendar_')]
@@ -62,12 +64,10 @@ class CalendarController extends AbstractApiController
     #[Route('', name: 'create', methods: Request::METHOD_POST)]
     public function create(#[CurrentUser] User $user, CreateCalendarRequest $request): JsonResponse
     {
-        $calendar = (new Calendar($request->getName()))
-            ->addUser($user)
-        ;
+        $calendar = new Calendar($request->getName(), $user);
 
-        foreach ($request->getUsers() as $user) {
-            $calendar->addUser($user);
+        foreach ($request->getCollaborators() as $collaborator) {
+            $calendar->addCollaborator($collaborator);
         }
 
         $this->calendarRepository->save($calendar);
@@ -76,15 +76,16 @@ class CalendarController extends AbstractApiController
     }
 
     #[Route('/{calendar}', name: 'update', methods: Request::METHOD_PUT)]
-    public function update(#[CurrentUser] User $user, Calendar $calendar, CreateCalendarRequest $request): JsonResponse
+    public function update(#[CurrentUser] User $user, Calendar $calendar, UpdateCalendarRequest $request): JsonResponse
     {
-        $calendar
-            ->setName($request->getName())
-            ->addUser($user)
-        ;
+        if ($calendar->getOwner() !== $user) {
+            throw new AccessDeniedException();
+        }
 
-        foreach ($request->getUsers() as $user) {
-            $calendar->addUser($user);
+        $calendar->setName($request->getName());
+
+        foreach ($request->getCollaborators() as $collaborator) {
+            $calendar->addCollaborator($collaborator);
         }
 
         $this->calendarRepository->save($calendar);
@@ -93,8 +94,12 @@ class CalendarController extends AbstractApiController
     }
 
     #[Route('/{calendar}', name: 'remove', methods: Request::METHOD_DELETE)]
-    public function remove(Calendar $calendar): JsonResponse
+    public function remove(#[CurrentUser] User $user, Calendar $calendar): JsonResponse
     {
+        if ($calendar->getOwner() !== $user) {
+            throw new AccessDeniedException();
+        }
+
         $this->calendarRepository->remove($calendar);
 
         return $this->respond($this->calendarRepository->findAll());
